requirements.txt
aiosqlite==0.19.0
discord.py==2.3.2
GitPython==3.1.40
icecream==2.1.3
numpy==1.26.2
Pillow==10.1.0
python-dotenv==1.0.0
Requests==2.31.0
